+++ testing/scie.py
@@ -5,24 +5,38 @@ from __future__ import absolute_import
 
 import pytest
 
+from pex.scie import Provider
 from pex.sysconfig import SysPlatform
+from pex.typing import TYPE_CHECKING
 from testing import IS_PYPY, PY_VER
+
+if TYPE_CHECKING:
+    from typing import Optional
+
+
+def provider():
+    # type: () -> Optional[Provider.Value]
+    if IS_PYPY:
+        if PY_VER == (2, 7):
+            return Provider.PyPy
+
+        if SysPlatform.LINUX_AARCH64 is SysPlatform.CURRENT and PY_VER >= (3, 7):
+            return Provider.PyPy
+        elif SysPlatform.MACOS_AARCH64 is SysPlatform.CURRENT and PY_VER >= (3, 8):
+            return Provider.PyPy
+        elif PY_VER >= (3, 6):
+            return Provider.PyPy
+        else:
+            return None
+    elif (3, 9) <= PY_VER < (3, 14):
+        return Provider.PythonBuildStandalone
+    else:
+        return None
 
 
 def has_provider():
     # type: () -> bool
-    if IS_PYPY:
-        if PY_VER == (2, 7):
-            return True
-
-        if SysPlatform.LINUX_AARCH64 is SysPlatform.CURRENT:
-            return PY_VER >= (3, 7)
-        elif SysPlatform.MACOS_AARCH64 is SysPlatform.CURRENT:
-            return PY_VER >= (3, 8)
-        else:
-            return PY_VER >= (3, 6)
-    else:
-        return (3, 9) <= PY_VER < (3, 14)
+    return provider() is not None
 
 
 skip_if_no_provider = pytest.mark.skipif(