+++ tests/integration/scie/test_pex_scie.py
@@ -315,7 +315,7 @@ def test_specified_science_binary(tmpdir):
 
     local_science_binary = os.path.join(str(tmpdir), "science")
     with open(local_science_binary, "wb") as write_fp, URLFetcher().get_body_stream(
-        "https://github.com/a-scie/lift/releases/download/v0.12.2/{binary}".format(
+        "https://github.com/a-scie/lift/releases/download/v0.12.4/{binary}".format(
             binary=SysPlatform.CURRENT.qualified_binary_name("science")
         )
     ) as read_fp:
@@ -359,7 +359,7 @@ def test_specified_science_binary(tmpdir):
         cached_science_binaries
     ), "Expected the local science binary to be used but not cached."
     assert (
-        "0.12.2"
+        "0.12.4"
         == subprocess.check_output(args=[local_science_binary, "--version"]).decode("utf-8").strip()
     )